+++ src/main/kotlin/se/kjellstrand/lsystem/model/LSTriple.kt
@@ -1,7 +1,7 @@
 package se.kjellstrand.lsystem.model
 
 data class LSTriple(
-    var x: Float,
-    var y: Float,
-    var w: Float
+    var x: Double,
+    var y: Double,
+    var w: Double
 )
